fix: default current year is selected in holiday table, plus moved the sort by drop down related code to a new component
